:recycle: dispatch store created and autologin
