+++ src/store/index.js
@@ -23,6 +23,9 @@ if (!env.test) {
 store.subscribe(throttle(savedAccountsSubscriber.bind(null, store), 1000));
 store.subscribe(throttle(followedAccountsSubscriber.bind(null, store), 1000));
 
+
+store.dispatch({ type: actionTypes.storeCreated });
+
 // ignore this in coverage as it is hard to test and does not run in production
 /* istanbul ignore if */
 if (module.hot) {
