+++ src/store/middlewares/login.js
@@ -1,33 +1,18 @@
 import i18next from 'i18next';
 import { getAccount } from '../../utils/api/account';
-import { getAutoLogInData, shouldAutoLogIn } from '../../utils/login';
+
 import { extractAddress, extractPublicKey } from '../../utils/account';
 import { accountLoggedIn, accountLoading, accountLoggedOut } from '../../actions/account';
-import { activePeerSet, activePeerUpdate } from '../../actions/peers';
-import networks from '../../constants/networks';
+
 import actionTypes from '../../constants/actions';
 import accountConfig from '../../constants/account';
-import settings from '../../constants/settings';
+
 import { errorToastDisplayed } from '../../actions/toaster';
 
 const { lockDuration } = accountConfig;
-const autoLogInIfNecessary = (store) => {
-  const autologinData = getAutoLogInData();
-  if (shouldAutoLogIn(autologinData)) {
-    store.dispatch(activePeerSet({
-      passphrase: autologinData[settings.keys.autologinKey],
-      network: { ...networks.customNode, address: autologinData[settings.keys.autologinUrl] },
-    }));
-    store.dispatch(activePeerUpdate({
-      online: true,
-    }));
-  }
-};
-
 const loginMiddleware = store => next => (action) => {
   if (action.type !== actionTypes.activePeerSet ||
       (!action.data.publicKey && !action.data.passphrase)) {
-    autoLogInIfNecessary();
     return next(action);
   }
   next(action);
